# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GaskeunKerja for Business \u2014 MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemOut"
                  },
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{job_id}/candidates": {
      "get": {
        "tags": [
          "matching"
        ],
        "summary": "Get Ranked Candidates",
        "description": "Status pill derivation (Area 1 T5, resolved 2026-07-12): 'Belum diundang' /\n'Menunggu wawancara' / 'Selesai wawancara' are derived from row presence\n(invited_at, interview_answers, hr_decisions), never a stored status field.",
        "operationId": "get_ranked_candidates_jobs__job_id__candidates_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchOut"
                  },
                  "title": "Response Get Ranked Candidates Jobs  Job Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/{candidate_id}/cv": {
      "get": {
        "tags": [
          "candidate_detail"
        ],
        "summary": "Get Candidate Cv",
        "description": "Streams the candidate's original uploaded CV PDF (Round-3 Task 18) \u2014 direct mirror of\nget_answer_audio below. HR-scoped like every other candidate-detail endpoint. The raw PDF\ncontains real PII (name/email/phone) by design \u2014 this is an HR-facing view, not redacted,\nsame as the existing raw_cv_path storage already was.",
        "operationId": "get_candidate_cv_candidates__candidate_id__cv_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetencyStatus": {
        "properties": {
          "competency_name": {
            "type": "string",
            "title": "Competency Name"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "proficiency": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proficiency"
          }
        },
        "type": "object",
        "required": [
          "competency_name",
          "matched",
          "proficiency"
        ],
        "title": "CompetencyStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          }
        },
        "type": "object",
        "required": [
          "title",
          "responsibilities",
          "requirements",
          "qualifications"
        ],
        "title": "JobCreateRequest"
      },
...
      "JobListItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          },
          "required_education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Education Level"
          },
          "required_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Major"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "question_status": {
            "type": "string",
            "title": "Question Status"
          },
          "belum_diundang": {
            "type": "integer",
            "title": "Belum Diundang"
          },
          "menunggu_wawancara": {
            "type": "integer",
            "title": "Menunggu Wawancara"
          },
          "selesai_wawancara": {
            "type": "integer",
            "title": "Selesai Wawancara"
          },
          "diputuskan": {
            "type": "integer",
            "title": "Diputuskan"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "title",
          "responsibilities",
          "requirements",
          "qualifications",
          "required_education_level",
          "required_major",
          "status",
          "created_at",
          "candidate_count",
          "question_status",
          "belum_diundang",
          "menunggu_wawancara",
          "selesai_wawancara",
          "diputuskan"
        ],
        "title": "JobListItemOut"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          },
          "required_education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Education Level"
          },
          "required_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Major"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "title",
          "responsibilities",
          "requirements",
          "qualifications",
          "required_education_level",
          "required_major",
          "status",
          "created_at"
        ],
        "title": "JobOut"
      },
...
      "MatchOut": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "alias": {
            "type": "string",
            "title": "Alias"
          },
          "overall_score": {
            "type": "number",
            "title": "Overall Score"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "competency_breakdown": {
            "type": "object",
            "title": "Competency Breakdown"
          },
          "competency_status": {
            "items": {
              "$ref": "#/components/schemas/CompetencyStatus"
            },
            "type": "array",
            "title": "Competency Status"
          },
          "latest_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Role"
          },
          "invited": {
            "type": "boolean",
            "title": "Invited"
          },
          "interview_completed": {
            "type": "boolean",
            "title": "Interview Completed"
          },
          "decided": {
            "type": "boolean",
            "title": "Decided"
          },
          "cv_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cv Url"
          },
          "education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education Level"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "meets_education": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meets Education"
          },
          "has_email": {
            "type": "boolean",
            "title": "Has Email"
          },
          "invite_email_sent": {
            "type": "boolean",
            "title": "Invite Email Sent"
          },
          "skill_gap_ready": {
            "type": "boolean",
            "title": "Skill Gap Ready"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "alias",
          "overall_score",
          "rank",
          "competency_breakdown",
          "competency_status",
          "latest_role",
          "invited",
          "interview_completed",
          "decided",
          "cv_url",
          "education_level",
          "major",
          "meets_education",
          "has_email",
          "invite_email_sent",
          "skill_gap_ready"
        ],
        "title": "MatchOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}